upgrade deps; rewrite smtp session

+++ smtp/manager.go
@@ -93,7 +93,7 @@ func NewManager(cfg *Config) *Manager {
 	s.ErrorLog = loggerWrapper{func(s string, i ...any) { cfg.Logger.Error().Msgf(s, i...) }}
 	s.ReadTimeout = 10 * time.Second
 	s.WriteTimeout = 10 * time.Second
-	s.MaxMessageBytes = cfg.MaxSize * 1024 * 1024
+	s.MaxMessageBytes = int64(cfg.MaxSize * 1024 * 1024)
 	s.AllowInsecureAuth = !cfg.TLSRequired
 	s.EnableREQUIRETLS = cfg.TLSRequired
 	s.EnableSMTPUTF8 = true
